youtube_summarizer_steam/requirements.txt
pytubefix
openai-whisper
reportlab
google-generativeai
